Move MeshCollector out of the way
