+++ src/wieldmesh.cpp
@@ -27,6 +27,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "mesh.h"
 #include "content_mapblock.h"
 #include "mapblock_mesh.h"
+#include "client/meshgen/collector.h"
 #include "client/tile.h"
 #include "log.h"
 #include "util/numeric.h"
